+++ tests/Unit/Workflow/ConfigWorkflowElementTest.php
@@ -5,13 +5,13 @@ namespace ArtARTs36\GitHandler\Tests\Unit\Workflow;
 use ArtARTs36\GitHandler\Config\Subjects\ConfigCommit;
 use ArtARTs36\GitHandler\Enum\ConfigSectionName;
 use ArtARTs36\GitHandler\Tests\Unit\GitTestCase;
-use ArtARTs36\GitHandler\Workflow\Elements\ConfigCommitWorkflowElement;
+use ArtARTs36\GitHandler\Backup\Elements\ConfigCommitWorkflowElement;
 
 final class ConfigWorkflowElementTest extends GitTestCase
 {
     /**
-     * @covers \ArtARTs36\GitHandler\Workflow\Elements\ConfigCommitWorkflowElement::dump
-     * @covers \ArtARTs36\GitHandler\Workflow\Elements\ConfigCommitWorkflowElement::__construct
+     * @covers \ArtARTs36\GitHandler\Backup\Elements\ConfigCommitWorkflowElement::dump
+     * @covers \ArtARTs36\GitHandler\Backup\Elements\ConfigCommitWorkflowElement::__construct
      */
     public function testDump(): void
     {
@@ -25,8 +25,8 @@ final class ConfigWorkflowElementTest extends GitTestCase
     }
 
     /**
-     * @covers \ArtARTs36\GitHandler\Workflow\Elements\ConfigCommitWorkflowElement::restore
-     * @covers \ArtARTs36\GitHandler\Workflow\Elements\ConfigCommitWorkflowElement::__construct
+     * @covers \ArtARTs36\GitHandler\Backup\Elements\ConfigCommitWorkflowElement::restore
+     * @covers \ArtARTs36\GitHandler\Backup\Elements\ConfigCommitWorkflowElement::__construct
      */
     public function testRestore(): void
     {
